src/test/java/com/jeremy7/sell/service/impl/categoryImplTest.java
package com.jeremy7.sell.service.impl;

import com.jeremy7.sell.dataobject.ProductCategory;
import org.junit.Assert;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.test.context.junit4.SpringRunner;

import static org.junit.Assert.*;
@RunWith(SpringRunner.class)
@SpringBootTest
public class categoryImplTest {

    @Autowired
    private  categoryImpl category;
    @Test
    public void findOne() {
      ProductCategory productCategory =  category.findOne(2);
        Assert.assertEquals(new Integer(2),productCategory.getCategoryId());
    }

    @Test
    public void findAll() {
    }

    @Test
    public void findByCategoryTypeIn() {
    }

    @Test
    public void save() {
    }
}